backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy[asyncio]
asyncpg
alembic
python-multipart
minio
chromadb
requests
pytest
pydantic
python-magic-bin
PyPDF2
sentence-transformers
passlib[bcrypt]
PyJWT
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy[asyncio]
asyncpg
alembic
python-multipart
minio
chromadb
requests
pydantic
python-magic-bin
PyPDF2
sentence-transformers
passlib[bcrypt]
PyJWT
